renaming api_notification.js to legacy.js for issue #2947
