+++ src/core/legacy.js
